+++ RELEASE-NOTES.md
@@ -1,3 +1,5 @@
 ### 1.0.0
 
 * Initial release
+* Added `#scite` parser function
+* Added `#referencelist` parser function